Create CalcZernikesBaseTask to separate out CalcZernikesTask dependency in calcZernikesNeuralTask.
